feat(cpu): math kernels (matmul/bmm/dot/pow/clamp) cover all dtypes and preserve F64 scalar precision
Audit Phase 2 Tasks 2.3+2.4+2.5+2.10:
- matmul: add Int16/Int64/UInt16/UInt32/UInt64 scalar paths with appropriate
  accumulator types. PyTorch-style wrap on overflow.
- bmm: when a dtype is in matmul's native list but not bmm's batched cblas
  list (F16/BF16/Complex/new ints), loop matmul_kernel over batch dim.
  cblas_*_batch_strided fast paths preserved for F32/F64.
- dot: add Float16/BFloat16 widen-narrow (F32 accumulator), Complex64/128
  std::complex accumulator, and the integer family.
- pow/clamp/clamp_min/clamp_max: widen scalar parameter from float to double
  through the kernel signatures and the registry lambdas, so Float64 callers
  no longer lose ~9 decimal digits of precision at the API boundary.

22/22 tests in tests/unit/test_math_dtype_coverage.cpp pass.

+++ include/tenzor/ops/math.hpp
@@ -113,7 +113,7 @@ auto baddbmm(const Tensor& input, const Tensor& batch1, const Tensor& batch2,
 /// @{
 
 /** @brief Element-wise power: input^exponent */
-auto pow(const Tensor& input, float exponent) -> Tensor;
+auto pow(const Tensor& input, double exponent) -> Tensor;
 
 /** @brief Element-wise exponential: e^input */
 auto exp(const Tensor& input) -> Tensor;
@@ -208,13 +208,13 @@ auto round(const Tensor& input) -> Tensor;
  * @param max Maximum value
  * @return Clamped tensor
  */
-auto clamp(const Tensor& input, float min, float max) -> Tensor;
+auto clamp(const Tensor& input, double min, double max) -> Tensor;
 
 /** @brief Clamp values to minimum */
-auto clamp_min(const Tensor& input, float min) -> Tensor;
+auto clamp_min(const Tensor& input, double min) -> Tensor;
 
 /** @brief Clamp values to maximum */
-auto clamp_max(const Tensor& input, float max) -> Tensor;
+auto clamp_max(const Tensor& input, double max) -> Tensor;
 
 /// @}
 /// @name Comparison Operations
